Split Logs.test.tsx to fix max-lines lint failure
Logs.test.tsx grew to 827 lines (539 counted, non-blank/non-comment)
after #71270 added copy/text-selection tests, tripping the max-lines
ESLint rule (cap 500) and failing the scheduled ARM CI's Static checks
job. Move the copy/selection describes and their helper functions into
a new LogsSelection.test.tsx. No test logic changed.
